dom/media/test/test_background_video_suspend.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test Background Video Suspends</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="manifest.js"></script>
<script src="background_video.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
<script type="text/javascript">
"use strict";

var manager = new MediaTestManager;

var MIN_DELAY = 100;

function testDelay(v, start, min) {
  let end = performance.now();
  let delay = end - start;
  ok(delay > min, `${v.token} suspended with a delay of ${delay} ms`);
}

startTest({
  desc: 'Test Background Video Suspends',
  prefs: [
    [ "media.test.video-suspend", true ],
    [ "media.suspend-bkgnd-video.enabled", true ],
    // User a short delay to ensure video decode suspend happens before end
    // of video.
    [ "media.suspend-bkgnd-video.delay-ms", MIN_DELAY ]
  ],
  tests: gDecodeSuspendTests,
  runTest: (test, token) => {
    let v = appendVideoToDoc(test.name, token);
    manager.started(token);

    let start;
    waitUntilPlaying(v)
      .then(() => { start = performance.now(); })
      .then(() => testVideoSuspendsWhenHidden(v))
      .then(() => {
        testDelay(v, start, MIN_DELAY);
        return testVideoResumesWhenShown(v);
      })
      .then(() => waitUntilEnded(v))
      .then(() => {
        removeNodeAndSource(v);
        manager.finished(token);
      });
  }
});
</script>
